% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{count_group}
\alias{count_group}
\title{Count how many items are in a group}
\usage{
count_group(dataset, ...)
}
\arguments{
\item{dataset}{dataframe}

\item{...}{character}
}
\value{
dataframe
}
\description{
Count how many items are in a group
}
