% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{identify_empty_columns}
\alias{identify_empty_columns}
\title{Find columns that are "empty" (only one unique variable)}
\usage{
identify_empty_columns(dataset)
}
\arguments{
\item{dataset}{dataframe}
}
\value{
character
}
\description{
Find columns that are "empty" (only one unique variable)
}
