% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{pull_count}
\alias{pull_count}
\title{Count how many items are in a group}
\usage{
pull_count(dataset)
}
\arguments{
\item{dataset}{dataframe}

\item{...}{character}
}
\value{
integer
}
\description{
Count how many items are in a group
}
