% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{remove_extra_columns}
\alias{remove_extra_columns}
\title{Remove columns that are "empty" (only one unique variable)}
\usage{
remove_extra_columns(dataset)
}
\arguments{
\item{dataset}{dataframe}
}
\value{
dataframe
}
\description{
Remove columns that are "empty" (only one unique variable)
}
